fix: Remove instanceof comparison for MongoDB DB instance check
We can not reasonably use an `instanceof` comparison to confirm that a
given object is of type MongoDB since we may be using a different
version of MongoDB than what a caller might be using.  In this
scenario the check would fail. Thus this patch checks if a required
field in the IDbProperties object is set. This required field is not a
known field in MongoDB instance, thus it is arguably sane to check it
in this manner.